Implementación de una estructura más organizada para el proyecto
