+++ PalacioVerduras.Domain/Entities/Metodos.cs
