+++ PalacioVerduras.Domain/Entities/ProductoVendido.cs
